# .eh_frame/parser: Only collect registers used for unwinding